requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.18
certifi==2017.11.5
chardet==3.0.4
diff-match-patch==20121119
Django==1.8.7
django-crispy-forms==1.7.2
django-decouple==2.1
django-extensions==2.0.7
-e git+https://github.com/django-import-export/django-import-export.git@657b7d72b36e61c3e2eaedd0ec8695bab638cf84#egg=django_import_export
django-jet==1.0.7
django-newsletter==0.6
django-parler==1.5.1
django-paypal==0.2.5
django-registration-redux==1.10
django-sendgrid-v5==0.6.87
django-tinymce==2.4.0
django-watson==1.5.2
django-webtest==1.9.2
djangorestframework==3.8.2
et-xmlfile==1.0.1
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
jdcal==1.4
kombu==3.0.37
ldif3==3.1.1
microsofttranslator==0.8
oauth2client==2.2.0
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.0.0
polib==1.1.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-card-me==0.9.3
python-dateutil==2.7.2
python-http-client==3.0.0
pytz==2017.3
PyYAML==3.12
redis==2.10.3
requests==2.18.4
rsa==3.4.2
sendgrid==5.3.0
six==1.11.0
sorl-thumbnail==12.4.1
South==1.0.2
surlex==0.2.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
waitress==1.1.0
WebOb==1.8.1
WebTest==2.0.29
xlrd==1.1.0
xlwt==1.3.0
